koji_cg: use listCGs method
Call listCGs before doing any grantCGAccess or revokeCGAccess calls.

This allows us to know what we *would* do ahead of time, so we can
support check_mode now also.

With this change, the koji_cg module now requires Koji 1.19 for
check_mode and properly reporting changes. If we don't have the listCGs
call on the hub, we will still fall back to the old behavior.