koji_tag: treat maxdepth empty strings as None
In some cases, it is not trivial to make Ansible set a variable to
"None" using Jinja filters. If the user does not enable Ansible's
"jinja2_native" setting, variables are often set to empty strings
instead of "None".

The koji_tag module passes the empty string directly to the
setInheritanceData RPC, which passes it into psychopg, which raises an
error on the hub.

There are ways to work around this: enable "jinja2_native" locally, or
use the "default(omit)" Jinja filter. These workarounds are complex and
surprising to new users.

For usability, treat empty maxdepth strings "" as none.

This new maxdepth-6.yml integration test triggers the psychopg error
without the corresponding koji_tag.py change.