koji_tag: non-empty maxdepth strings should be ints
In some cases, it is not trivial to make Ansible set a variable to
an integer using Jinja filters. If the user does not enable Ansible's
"jinja2_native" setting, variables are often set to strings instead of
ints.

The describe_inheritance() method always requires an integer, so it
crashes when we pass in a string instead.

Commit ddf233dcc7efa79d705da7c35d7a8a4e874403f6 makes empty strings
"None". This change builds on that to make non-empty strings integers.

The new maxdepth-7.yml integration test crashes in the format string in
describe_inheritance() without this change.

Thanks Caleigh Runge-Hottman <caleigh@redhat.com> for tracking down this
problem.