tests: fix missing listCGs method error message
The error message for any missing method looks like "Invalid method: foo".

Prior to this change, returned an incorrect string in our faked error in
the unit test.

We are not exercising the error message string comparison logic in
list_cgs() yet, so this bug did not result in a test failure. I'll add
test coverage for list_cgs() in a future commit.