koji_tag: inheritance input normalization
This change separates inheritance input normalization from parent/child
tag ID determination. It also adds a normalization process to priority
(always interpret as an integer).

Includes a unit test to check that string priority inputs function the
same as integer priority inputs, and an integration test to verify that
a string priority input will not be detected as a change.