tests: assert koji_tag.ensure_external_repos behavior
Prior to this change, we did not check that the ensure_external_repos()
method made any changes to our fixture.

Fix the getTagExternalRepos() and addExternalRepoToTag() methods in our
FakeSession fixture to store and return tag-repo associations correctly,
using a tag_repo dict.

Verify the behavior of the method by asserting the result of
getTagExternalRepos() from our FakeSession test fixture.