koji_tag: remove then add in ensure_external_repos
Prior to this change, we would hit an error in the hub because we would
try to add an external repo at the same priority as one that already
exists.

Switch the order of operations so that we delete all the wrong repos
from the hub and then add all the correct repos that are missing.

This change drops our use of editExternalRepos. It would be faster and
idempotent to use that RPC instead of the pair of remove+add RPCs,
however the editExternalRepos RPC silently resets merge_mode if the Koji
hub is older than v1.21 (https://pagure.io/koji/issue/1857). Maybe we
can find a better way to use editExternalRepos in the future.

This also moves our add/remove RPC loops to make those easier to test
and refactor to multicalls later.