remove the compound_parameter_present method from library.koji_tag
This method is not necessary if we use the correct types for the type
property of the argument spec. We still need to check to see if the
argument is given before operating against it though. The check for
None is important because if given an empty list or dict then we
want to remove.

Fixes #155

Signed-off-by: Andrew Schoen <aschoen@redhat.com>