koji_tag: gracefully handle empty strings for parameters
As described in ddf233dcc7efa79d705da7c35d7a8a4e874403f6, Ansible's
default jinja2_native setting is False, and when this is False, the
"default(None)" filter evaluates to an empty string.

With the changes in 8f9f1c1aa59e8463ca8425636073633650061266, we failed
to process parameters with empty strings.

Update the "inheritance" and "external_repos" checks to test for lists
of empty strings and skip them.

Update the "packages" and "groups" argspecs to "raw" in order to restore
the old behavior before 8f9f1c1aa59e8463ca8425636073633650061266.