koji_tag: do not require logins for check-mode
Prior to this change, if a user ran a koji_tag task that managed
packages or groups on a tag, we unconditionally called koji's
activate_session() method.

This is bad for performance because it puts load on the hub when we
might not need to authorize any changes. When users run check mode, we
should never authenticate. (This also means we ought to let users run
check mode without valid credentials.)