build-collection: fix VERSION for non-master tests
When a GitHub Action runner checks out a non-master branch for testing,
it checks out GitHub's magic "merge" ref. For example:

  git checkout --progress --force refs/remotes/pull/192/merge

This puts the repository into a detached HEAD state. This means that we
cannot search "HEAD" for "--max-parents=0", because Git will simply return
the sha we've checked out. This problem means that COMMIT_COUNT is
always "-1" for pull requests and VERSION is always "0.0.-1".

GitHub Action runners have a GITHUB_BASE_REF environment variable with
the name of a pull request base branch. If that is defined, use that to
discover the root commit of the base branch. This fixes the COMMIT_COUNT
calculation for pull requests.