remove owner argument to packageListRemove
packageListRemove does not take an "owner" argument. Remove owner from
the places we call packageListRemove.

The reason this code did not crash is that packageListRemove takes a
third argument "force". (Koji has had this packageListRemove parameter
since the beginning.) The hub evaluated the owner name as "True" in the
boolean context. As long as the Ansible user had admin permissions,
force="ownername" did not raise an error.