koji_tag_packages: blocked_packages fixes (#1)
* tag_packages: fix test assertion for packageListBlock

The packageListBlock RPC takes two arguments, not three.

* tag_packages: remove unused conditional

* tag_packages: handle listPackages() in hubs prior to v1.25

v1.25.0 added the with_owners kwarg for performance, but that release is
only five months old. Let's keep support for older hubs.

* tag_packages: login immediately before packageListBlock

Only login if we are not in check_mode and we have changes to make
(we're about to call packageListBlock).

* tag_packages: rm "koji: kojidev" from blocked pkg examples

Remove this setting to make the examples simpler.

* tag_packages: simplify current_blocked list comprehension

Simplify the boolean evaluation for pkg['blocked'].

* tag_packages: simplify blocked_packages implementation

Remove the "blocked" and "unblocked" states.

Make blocked_packages a simple list parameter of package names that we
will either "block" or "unblock" according to "state: present" or
"state: absent".

* tag_packages: add package block integration test

Assert that we can block and unblock a package from a child tag.

* tag_packages: only remove package blocks if present

Check listPackages to learn which packages to unblock, rather than
directly calling packageListUnblock on the entire list.

With this change, koji_tag_packages only reports the packages that
it's changing, rather than reporting potentially false changes every
time.