tag_packages: handle listPackages' with_owners arg in tests
listPackages in Koji v1.25+ takes a with_owners parameter. Accept this
parameter in the FakeKojiSession in the test suite.

The test suite currently mocks session.listPackages() with a Mock()
object rather than using FakeKojiSession(), so this does not affect the
test runs yet. However, it's important to make this more robust for when
we switch over to using FakeKojiSession().